+++ Classes/Code/TagManagerCodeBuilder.php
@@ -107,13 +107,15 @@ class TagManagerCodeBuilder
 
     private function addContainerCode(): void
     {
-        $this->codeParts[] = new JavaScriptCode(
-            'var d=document,g=d.createElement("script"),s=d.getElementsByTagName("script")[0];'
-            . \sprintf(
-                'g.async=true;g.src="%sjs/container_%s.js";s.parentNode.insertBefore(g,s);',
-                $this->configuration->url,
-                $this->configuration->tagManagerContainerId
-            )
-        );
+        foreach ($this->configuration->tagManagerContainerId as $containerId) {
+            $this->codeParts[] = new JavaScriptCode(
+                'var d=document,g=d.createElement("script"),s=d.getElementsByTagName("script")[0];'
+                . \sprintf(
+                    'g.async=true;g.src="%sjs/container_%s.js";s.parentNode.insertBefore(g,s);',
+                    $this->configuration->url,
+                    $containerId
+                )
+            );
+        }
     }
 }