+++ Tests/Unit/Code/TagManagerCodeBuilderTest.php
@@ -111,4 +111,25 @@ final class TagManagerCodeBuilderTest extends TestCase
             $subject->getCode()
         );
     }
+
+    /**
+     * @test
+     */
+    public function getCodeReturnsTagManagerCodeWithMultipleContainerIdsCorrectly(): void
+    {
+        $this->subject
+            ->setRequest($this->requestStub)
+            ->setConfiguration(
+                Configuration::createFromSiteConfiguration([
+                    'matomoIntegrationUrl' => 'https://www.example.net/',
+                    'matomoIntegrationSiteId' => 123,
+                    'matomoIntegrationTagManagerContainerId' => 'someId,anotherId',
+                ])
+            );
+
+        self::assertSame(
+            'var _mtm=window._mtm||[];_mtm.push({"mtm.startTime":(new Date().getTime()),"event":"mtm.Start"});var d=document,g=d.createElement("script"),s=d.getElementsByTagName("script")[0];g.async=true;g.src="https://www.example.net/js/container_someId.js";s.parentNode.insertBefore(g,s);var d=document,g=d.createElement("script"),s=d.getElementsByTagName("script")[0];g.async=true;g.src="https://www.example.net/js/container_anotherId.js";s.parentNode.insertBefore(g,s);',
+            $this->subject->getCode()
+        );
+    }
 }