+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Multiple tag manager container IDs can be used (#14)
+
 ## [1.5.1] - 2023-04-01
 
 ### Fixed